split out child packages
